LMAO IT WORKS NOW
okay so not only is everything commented now, but it all seems to work...

the reason for the previous issue?
me forgetting that I had left the debug section in the script
(it now is commented out).
yep, that's it. that's the entire reason.

sometimes I despise how dumb programming makes you feel.